Only save uuid file if the partition layout is GPT
git-svn-id: https://svn.code.sf.net/p/freeghost/code/trunk@4735 71f96598-fa45-0410-b640-bcd6f8691b32

+++ src/buildroot/package/fog/scripts/usr/share/fog/lib/partition-funcs.sh
@@ -25,6 +25,9 @@ saveUUIDInformation() {
     local file="$2"
     [[ -z $disk ]] && handleError "No disk passed (${FUNCNAME[0]})"
     [[ -z $file ]] && handleError "No file to save to passed (${FUNCNAME[0]})"
+    local hasgpt=0
+    hasGPT "$disk"
+    [[ $hasgpt -eq 0 ]] && return
     rm -f $file
     local diskuuid=""
     local partuuid=""
